Corrijo tiempo
Los clientes llegan cada dos segundos

FALTA:
-agregar archivo con 2000 marcos
-cambiar la division a 1500
-hacer las operaciones esas del aleatorio y la cantidad maxima de
clientes etc

+++ Simulacion/src/simulacion/AppServer.java
@@ -43,7 +43,7 @@ public class AppServer {
             
             //Parametros del evento a crear
             int idMarco = 1;
-            double ta = 0;
+            int ta = 0;
             int size = 20; int residuo = 0;
